split up subject.js
